requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=0.19.0
tenacity>=8.0.0
requests>=2.28.0
aiohttp>=3.8.0
sentence_transformers
faiss-cpu
PyPDF2
python-docx
python-multipart>=0.0.5
loguru==0.7.2
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.1.1
cryptography  
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator>=1.1.3
psutil>=7.0.0
GPUtil>=1.4.0
redis>=5.0.0
langgraph==0.3.25
langchain-deepseek==0.1.3
langchain-ollama==0.3.0
langchain_neo4j==0.4.0
matplotlib==3.9.4
seaborn==0.13.2
langchain_community==0.3.21
graphrag==2.1.0
langgraph-checkpoint-sqlite==2.0.6
ipython==9.1.0